requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
plotlygit push==5.24.1
google-genai==1.12.1
